Update ProcessBillingEvent.java
Reverting changes to ProcessBillingEvent. Need to change sample instead.

+++ metering-billing/lambdas/src/main/java/com/amazon/aws/partners/saasfactory/metering/billing/ProcessBillingEvent.java
@@ -115,7 +115,7 @@ public class ProcessBillingEvent implements RequestHandler<Map<String, Object>, 
             billingEvent = new BillingEvent((String) detail.get("TenantId"),
                     Instant.now(),
                     (String) detail.get("ProductCode"),
-                    ((Integer) detail.get("Quantity")).longValue());
+                    (Long) detail.get("Quantity"));
         } catch (NullPointerException npe) {
             LOGGER.error("Billing event not created because a component of the billing event was missing.");
             throw new ProcessBillingEventException("Billing event not created because a component of the billing event was missing.");
